Add `EvalPool.alignment()` method.
This returns a set of outcomes that should always be seen by `next_state` even if they have 0 count.

If all outcomes are `int`s, the default implementation goes through enough consecutive integers to cover the range between the min and max possible outcome inclusive. I'm not sure I like this though, so this is likely to change in the future.